+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.10.6] - 2026-01-10
 
+### Added
+- Focus Sessions - block all sites in a folder with one shared timer
+- Focus button on folder headers opens modal to set duration
+- Quick duration presets (15m, 30m, 1h, 2h) in focus modal
+- Focus sessions override individual site settings - blocks everything in the folder
+
 ### Changed
 - Timer blocking now works as temporary block - sites are accessible until you start the timer
 - Timer sites show "Disabled" button that starts blocking when clicked, "Stop" to end early